Добавила удаление пользователя по id, откорректировала методы rest-контроллера, добавила обработку исключений

+++ src/main/java/ru/kata/spring/boot_security/demo/controllers/AdminController.java
@@ -3,9 +3,16 @@ package ru.kata.spring.boot_security.demo.controllers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import ru.kata.spring.boot_security.demo.exception_handling.NoSuchUserException;
 import ru.kata.spring.boot_security.demo.models.User;
-import ru.kata.spring.boot_security.demo.services.RoleService;
 import ru.kata.spring.boot_security.demo.services.UserService;
 
 import java.util.List;
@@ -14,12 +21,10 @@ import java.util.List;
 @RequestMapping("/admin")
 public class AdminController {
     private final UserService userService;
-    private final RoleService roleService;
 
     @Autowired
-    public AdminController(UserService userService, RoleService roleService) {
+    public AdminController(UserService userService) {
         this.userService = userService;
-        this.roleService = roleService;
     }
 
     @GetMapping("")
@@ -29,47 +34,43 @@ public class AdminController {
         return new ResponseEntity<>(users, HttpStatus.OK);
     }
 
-    @GetMapping("/create")
-    public ResponseEntity<User> create() {
-        User user = new User();
-        // Метод create() теперь возвращает ResponseEntity с объектом User
-        return new ResponseEntity<>(user, HttpStatus.OK);
+    @GetMapping("/{id}")
+    public User getUserById(@PathVariable long id) {
+        User user = userService.getById(id);
+
+        if(user == null) {
+            throw new NoSuchUserException("Нет пользователя с ID = " +
+                    id + " в Базе Данных");
+        }
+
+        return user; //возвращается НЕ сам объект user, а JSON, соответствующий этому объекту
     }
 
     @PostMapping("/create")
     // Аннотация @ModelAttribute заменена на @RequestBody
-    public ResponseEntity<Void> addUser(@RequestBody User user) {
+    public ResponseEntity<User> addUser(@RequestBody User user) {
         userService.save(user);
         // Метод addUser() теперь возвращает ResponseEntity с HttpStatus.CREATED
-        return new ResponseEntity<>(HttpStatus.CREATED);
+        return new ResponseEntity<>(user, HttpStatus.CREATED);
     }
 
-    @GetMapping("/edit")
-    public ResponseEntity<User> edit(@RequestParam(value = "id", required = false) Long id) {
+    // Put метод с возвращением НЕ самого объекта user, а JSON, соответствующий обновленному пользователю
+    @PutMapping("/edit")
+    public User edit(@RequestBody User user) {
+        userService.update(user);
+        return user;
+    }
+
+    @DeleteMapping("/{id}")
+    public String deleteUser (@PathVariable long id) {
         User user = userService.getById(id);
-        // Метод edit() теперь возвращает ResponseEntity с объектом User
-        return new ResponseEntity<>(user, HttpStatus.OK);
-    }
 
-    @PostMapping("/edit")
-    // Аннотация @ModelAttribute заменена на @RequestBody
-    public ResponseEntity<Void> update(@RequestBody User userForm) {
-        userService.update(userForm);
-        // Метод update() теперь возвращает ResponseEntity с HttpStatus.NO_CONTENT (возвращает статус 204 (No Content) для обозначения того, что обновление было успешным)
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-    }
-
-    // вариант с возвращением обновленного пользователя вместе с статусом 200 (OK) для обозначения того, что обновление было успешным и мы возвращаем обновленного пользователя
-//    @PostMapping("/edit")
-//public ResponseEntity<User> update(@RequestBody User userForm) {
-//    userService.update(userForm);
-//    return new ResponseEntity<>(userForm, HttpStatus.OK);
-//
-
-    @PostMapping("/delete")
-    public ResponseEntity<Void> delete(@RequestParam(value = "id") Long id) {
+        if (user == null) {
+            throw new NoSuchUserException("Пользователя с ID = " +
+                    id + " не существует в Базе Данных");
+        }
         userService.deleteById(id);
-        // Метод delete() теперь возвращает ResponseEntity с HttpStatus.NO_CONTENT
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+
+        return "Пользователь с ID = " + id + " был успешно удалён";
     }
 }
